Rename org.keycloak.test to org.keycloak.tests
Closes #36389
