--- tests/base/src/test/java/org/keycloak/test/admin/authz/fgap/FeatureV1EnabledTest.java
@@ -1,40 +1,0 @@
-/*
- * Copyright 2025 Red Hat, Inc. and/or its affiliates
- * and other contributors as indicated by the @author tags.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package org.keycloak.test.admin.authz.fgap;
-
-import static org.hamcrest.MatcherAssert.assertThat;
-import static org.hamcrest.Matchers.notNullValue;
-import static org.hamcrest.Matchers.nullValue;
-import org.junit.jupiter.api.Test;
-import org.keycloak.representations.idm.authorization.ResourceServerRepresentation;
-import org.keycloak.testframework.annotations.InjectClient;
-import org.keycloak.testframework.annotations.KeycloakIntegrationTest;
-import org.keycloak.testframework.realm.ManagedClient;
-
-@KeycloakIntegrationTest(config = KeycloakAdminPermissionsV1ServerConfig.class)
-public class FeatureV1EnabledTest {
-
-    @InjectClient(ref = "test-client", config = AuthzClientConfig.class, createClient = true)
-    private ManagedClient testClient;
-
-    @Test
-    public void schemaNotAvailableFeatureV1Enabled() {
-        ResourceServerRepresentation authorizationSettings = testClient.admin().authorization().getSettings();
-        assertThat(authorizationSettings, notNullValue());
-        assertThat(authorizationSettings.getAuthorizationSchema(), nullValue());
-    }
-}
