--- tests/base/src/test/java/org/keycloak/test/suites/DatabaseTestSuite.java
@@ -1,9 +1,0 @@
-package org.keycloak.test.suites;
-
-import org.junit.platform.suite.api.SelectPackages;
-import org.junit.platform.suite.api.Suite;
-
-@Suite
-@SelectPackages({"org.keycloak.test.admin"})
-public class DatabaseTestSuite {
-}
