+++ tests/base/src/test/java/org/keycloak/tests/suites/DatabaseTestSuite.java
@@ -1,0 +1,9 @@
+package org.keycloak.tests.suites;
+
+import org.junit.platform.suite.api.SelectPackages;
+import org.junit.platform.suite.api.Suite;
+
+@Suite
+@SelectPackages({"org.keycloak.tests.admin"})
+public class DatabaseTestSuite {
+}
