+++ tests/base/src/test/resources/keycloak-test.properties
@@ -2,7 +2,7 @@ kc.test.log.level=WARN
 
 kc.test.log.filter=true
 
-kc.test.log.category."org.keycloak.test".level=INFO
+kc.test.log.category."org.keycloak.tests".level=INFO
 
 kc.test.log.category."testinfo".level=INFO
 kc.test.log.category."org.keycloak".level=WARN
