+++ tests/base/src/test/resources/org/keycloak/tests/admin/client/KEYCLOAK-4040-sharefile-metadata.xml
